--- app/src/main/java/com/example/immersioncamp1/CustomGalleryActivity.java
@@ -1,200 +1,0 @@
-package com.example.immersioncamp1;
-
-import android.Manifest;
-import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.database.Cursor;
-import android.net.Uri;
-import android.os.Build;
-import android.provider.MediaStore;
-
-import androidx.annotation.RequiresApi;
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-import androidx.core.app.ActivityCompat;
-import androidx.core.content.ContextCompat;
-import androidx.recyclerview.widget.RecyclerView;
-
-import android.util.Log;
-import android.view.View;
-import android.view.Window;
-import android.view.WindowManager;
-import android.widget.ImageButton;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import com.example.immersioncamp1.utils.ImageFolder;
-import com.example.immersioncamp1.utils.ItemClickListener;
-import com.example.immersioncamp1.utils.MarginDecoration;
-import com.example.immersioncamp1.utils.PicHolder;
-import com.example.immersioncamp1.utils.PictureFacer;
-import com.example.immersioncamp1.utils.PictureFolderAdapter;
-
-import java.util.ArrayList;
-
-public class CustomGalleryActivity extends AppCompatActivity implements ItemClickListener {
-
-    private final String TAG = "+CustomGalleryActivity";
-    RecyclerView folderRecycler;
-    TextView empty;
-    private static final int MY_PERMISSIONS_REQUEST_READ_EXTERNAL_STORAGE = 1;
-
-    /**
-     * Request the user for permission to access media files and read images on the device
-     * this will be useful as from api 21 and above, if this check is not done the Activity will crash
-     * <p>
-     * Setting up the RecyclerView and getting all folders that contain pictures from the device
-     * the getPicturePaths() returns an ArrayList of imageFolder objects that is then used to
-     * create a RecyclerView Adapter that is set to the RecyclerView
-     *
-     * @param savedInstanceState saving the activity state
-     */
-    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_custom_gallery);
-
-        if (ContextCompat.checkSelfPermission(CustomGalleryActivity.this,
-                Manifest.permission.READ_EXTERNAL_STORAGE)
-                != PackageManager.PERMISSION_GRANTED)
-            ActivityCompat.requestPermissions(CustomGalleryActivity.this,
-                    new String[]{Manifest.permission.READ_EXTERNAL_STORAGE},
-                    MY_PERMISSIONS_REQUEST_READ_EXTERNAL_STORAGE);
-        //____________________________________________________________________________________
-
-        empty = findViewById(R.id.empty);
-
-        folderRecycler = findViewById(R.id.folderRecycler);
-        folderRecycler.addItemDecoration(new MarginDecoration(CustomGalleryActivity.this));
-        folderRecycler.hasFixedSize();
-        ArrayList<ImageFolder> folds = getPicturePaths();
-
-        if (folds.isEmpty()) {
-            empty.setVisibility(View.VISIBLE);
-        } else {
-            RecyclerView.Adapter<PictureFolderAdapter.FolderHolder> folderAdapter = new PictureFolderAdapter(folds, CustomGalleryActivity.this, this);
-            folderRecycler.setAdapter(folderAdapter);
-        }
-
-        changeStatusBarColor();
-    }
-
-    /**
-     * 1
-     *
-     * @return gets all folders with pictures on the device and loads each of them in a custom object imageFolder
-     * the returns an ArrayList of these custom objects
-     */
-    private ArrayList<ImageFolder> getPicturePaths() {
-        ArrayList<ImageFolder> picFolders = new ArrayList<>();
-        ArrayList<String> picPaths = new ArrayList<>();
-        Uri allImagesUri = MediaStore.Images.Media.EXTERNAL_CONTENT_URI;
-        String[] projection = {
-                MediaStore.Images.Media.DISPLAY_NAME,
-                MediaStore.Images.ImageColumns.DATA,
-                MediaStore.Images.Media.BUCKET_DISPLAY_NAME,
-                MediaStore.Images.Media.BUCKET_ID};
-        Cursor cursor = this.getContentResolver().query(allImagesUri, projection, null, null, null);
-        try {
-            if (cursor != null) {
-                cursor.moveToFirst();
-            }
-            do {
-                ImageFolder folds = new ImageFolder();
-                String name = cursor.getString(cursor.getColumnIndexOrThrow(MediaStore.Images.Media.DISPLAY_NAME));
-                String folder = cursor.getString(cursor.getColumnIndexOrThrow(MediaStore.Images.Media.BUCKET_DISPLAY_NAME));
-                String datapath = cursor.getString(cursor.getColumnIndexOrThrow(MediaStore.Images.Media.DATA));
-
-                //String folderpaths =  datapath.replace(name,"");
-                String folderpaths = datapath.substring(0, datapath.lastIndexOf(folder + "/"));
-                folderpaths = folderpaths + folder + "/";
-                if (!picPaths.contains(folderpaths)) {
-                    picPaths.add(folderpaths);
-
-                    folds.setPath(folderpaths);
-                    folds.setFolderName(folder);
-                    folds.setFirstPic(datapath);//if the folder has only one picture this line helps to set it as first so as to avoid blank image in itemview
-                    folds.addpics();
-                    picFolders.add(folds);
-                } else {
-                    for (int i = 0; i < picFolders.size(); i++) {
-                        if (picFolders.get(i).getPath().equals(folderpaths)) {
-                            picFolders.get(i).setFirstPic(datapath);
-                            picFolders.get(i).addpics();
-                        }
-                    }
-                }
-            } while (cursor.moveToNext());
-            cursor.close();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        for (int i = 0; i < picFolders.size(); i++) {
-            Log.d("picture folders", picFolders.get(i).getFolderName() + " and path = " + picFolders.get(i).getPath() + " " + picFolders.get(i).getNumberOfPics());
-        }
-
-        //reverse order ArrayList
-       /* ArrayList<imageFolder> reverseFolders = new ArrayList<>();
-
-        for(int i = picFolders.size()-1;i > reverseFolders.size()-1;i--){
-            reverseFolders.add(picFolders.get(i));
-        }*/
-
-        return picFolders;
-    }
-
-
-    @Override
-    public void onPicClicked(PicHolder holder, int position, ArrayList<PictureFacer> pics) {
-
-    }
-
-    /**
-     * Each time an item in the RecyclerView is clicked this method from the implementation of the transitListerner
-     * in this activity is executed, this is possible because this class is passed as a parameter in the creation
-     * of the RecyclerView's Adapter, see the adapter class to understand better what is happening here
-     *
-     * @param pictureFolderPath a String corresponding to a folder path on the device external storage
-     */
-    @Override
-    public void onPicClicked(String pictureFolderPath, String folderName) {
-        Intent move = new Intent(CustomGalleryActivity.this, ImageDisplayActivity.class);
-        move.putExtra("folderPath", pictureFolderPath);
-        move.putExtra("folderName", folderName);
-
-        //move.putExtra("recyclerItemSize",getCardsOptimalWidth(4));
-        startActivity(move);
-    }
-
-
-    /**
-     * Default status bar height 24dp,with code API level 24
-     */
-    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
-    private void changeStatusBarColor() {
-        Window window = this.getWindow();
-        window.clearFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS);
-        window.addFlags(WindowManager.LayoutParams.FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS);
-        window.setStatusBarColor(ContextCompat.getColor(getApplicationContext(), R.color.black));
-
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-
-        Log.e(TAG, "onResume");
-    }
-
-    @Override
-    protected void onRestart() {
-        super.onRestart();
-        Log.e(TAG, "onRestart");
-        if (ContactDetailActivity.getImgPathViewModel().getImgPath() != null) {
-            finish();
-        }
-    }
-}
