--- app/src/main/java/com/example/immersioncamp1/utils/AutoFitRecyclerView.java
@@ -1,68 +1,0 @@
-package com.example.immersioncamp1.utils;
-
-import android.content.Context;
-import android.content.res.TypedArray;
-import android.util.AttributeSet;
-
-import androidx.recyclerview.widget.GridLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
-
-/**
- * Author CodeBoy722
- * A custom RecyclerView with GridLayout AutoFit Behavior
- */
-public class AutoFitRecyclerView extends RecyclerView {
-
-    private int columnWidth = -1;
-    private GridLayoutManager manager;
-
-
-    public AutoFitRecyclerView(Context context) {
-        super(context);
-        initialize(context, null);
-    }
-
-    public AutoFitRecyclerView(Context context, AttributeSet attrs) {
-        super(context, attrs);
-        initialize(context, attrs);
-    }
-
-    public AutoFitRecyclerView(Context context, AttributeSet attrs, int defStyle) {
-        super(context, attrs, defStyle);
-        initialize(context, attrs);
-    }
-
-
-    /**
-     *
-     * @param context The Activities Context
-     * @param attrs Dimention columnWidth of the RecyclerView
-     */
-    private void initialize(Context context, AttributeSet attrs) {
-        if (attrs != null) {
-            int[] attrsArray = {
-                    android.R.attr.columnWidth
-            };
-            TypedArray array = context.obtainStyledAttributes(attrs, attrsArray);
-            columnWidth = array.getDimensionPixelSize(0, -1);
-            array.recycle();
-        }
-
-        manager = new GridLayoutManager(getContext(), 1);
-        setLayoutManager(manager);
-    }
-
-
-    @Override
-    protected void onMeasure(int widthSpec, int heightSpec) {
-        super.onMeasure(widthSpec, heightSpec);
-        if (columnWidth > 0) {
-            int spanCount = Math.max(1, getMeasuredWidth() / columnWidth);
-            manager.setSpanCount(spanCount);
-        }
-    }
-
-
-}
-
-
