--- app/src/main/java/com/example/immersioncamp1/utils/ImageFolder.java
@@ -1,61 +1,0 @@
-package com.example.immersioncamp1.utils;
-
-/**
- * author CodeBoy722
- *
- * Custom Class that holds information of a folder containing images
- * on the device external storage, used to populate our RecyclerView of
- * picture folders
- */
-public class ImageFolder {
-
-    private  String path;
-    private  String FolderName;
-    private int numberOfPics = 0;
-    private String firstPic;
-
-    public ImageFolder(){
-
-    }
-
-    public ImageFolder(String path, String folderName) {
-        this.path = path;
-        FolderName = folderName;
-    }
-
-    public String getPath() {
-        return path;
-    }
-
-    public void setPath(String path) {
-        this.path = path;
-    }
-
-    public String getFolderName() {
-        return FolderName;
-    }
-
-    public void setFolderName(String folderName) {
-        FolderName = folderName;
-    }
-
-    public int getNumberOfPics() {
-        return numberOfPics;
-    }
-
-    public void setNumberOfPics(int numberOfPics) {
-        this.numberOfPics = numberOfPics;
-    }
-
-    public void addpics(){
-        this.numberOfPics++;
-    }
-
-    public String getFirstPic() {
-        return firstPic;
-    }
-
-    public void setFirstPic(String firstPic) {
-        this.firstPic = firstPic;
-    }
-}
