--- app/src/main/java/com/example/immersioncamp1/utils/PictureAdapter.java
@@ -1,76 +1,0 @@
-package com.example.immersioncamp1.utils;
-
-import static androidx.core.view.ViewCompat.setTransitionName;
-
-import android.content.Context;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
-
-import com.example.immersioncamp1.R;
-import com.bumptech.glide.Glide;
-import com.bumptech.glide.request.RequestOptions;
-
-import java.util.ArrayList;
-
-/**
- * Author CodeBoy722
- *
- * A RecyclerView Adapter class that's populates a RecyclerView with images from
- * a folder on the device external storage
- */
-public class PictureAdapter extends RecyclerView.Adapter<PicHolder> {
-
-    private ArrayList<PictureFacer> pictureList;
-    private Context pictureContx;
-    private final ItemClickListener picListerner;
-
-    /**
-     *
-     * @param pictureList ArrayList of pictureFacer objects
-     * @param pictureContx The Activities Context
-     * @param picListerner An interface for listening to clicks on the RecyclerView's items
-     */
-    public PictureAdapter(ArrayList<PictureFacer> pictureList, Context pictureContx, ItemClickListener picListerner) {
-        this.pictureList = pictureList;
-        this.pictureContx = pictureContx;
-        this.picListerner = picListerner;
-    }
-
-    @NonNull
-    @Override
-    public PicHolder onCreateViewHolder(@NonNull ViewGroup container, int position) {
-        LayoutInflater inflater = LayoutInflater.from(container.getContext());
-        View cell = inflater.inflate(R.layout.pic_holder_item, container, false);
-        return new PicHolder(cell);
-    }
-
-    @Override
-    public void onBindViewHolder(@NonNull final PicHolder holder, final int position) {
-
-        final PictureFacer image = pictureList.get(position);
-
-        Glide.with(pictureContx)
-                .load(image.getPicturePath())
-                .apply(new RequestOptions().centerCrop())
-                .into(holder.picture);
-
-        setTransitionName(holder.picture, String.valueOf(position) + "_image");
-
-        holder.picture.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                picListerner.onPicClicked(holder,position, pictureList);
-            }
-        });
-
-    }
-
-    @Override
-    public int getItemCount() {
-        return pictureList.size();
-    }
-}
