--- app/src/main/java/com/example/immersioncamp1/utils/RecyclerViewPagerImageIndicator.java
@@ -1,76 +1,0 @@
-package com.example.immersioncamp1.utils;
-
-import android.content.Context;
-import android.graphics.Color;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
-
-import com.example.immersioncamp1.R;
-import com.bumptech.glide.Glide;
-import com.bumptech.glide.request.RequestOptions;
-
-import java.util.ArrayList;
-
-/**
- * Author CodeBoy722
- */
-public class RecyclerViewPagerImageIndicator extends RecyclerView.Adapter<IndicatorHolder> {
-
-    ArrayList<PictureFacer> pictureList;
-    Context pictureContx;
-    private final ImageIndicatorListener imageListerner;
-
-    /**
-     *
-     * @param pictureList ArrayList of pictureFacer objects
-     * @param pictureContx The Activity of fragment context
-     * @param imageListerner Interface for communication between adapter and fragment
-     */
-    public RecyclerViewPagerImageIndicator(ArrayList<PictureFacer> pictureList, Context pictureContx, ImageIndicatorListener imageListerner) {
-        this.pictureList = pictureList;
-        this.pictureContx = pictureContx;
-        this.imageListerner = imageListerner;
-    }
-
-
-    @NonNull
-    @Override
-    public IndicatorHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        LayoutInflater inflater = LayoutInflater.from(parent.getContext());
-        View cell = inflater.inflate(R.layout.indicator_holder, parent, false);
-        return new IndicatorHolder(cell);
-    }
-
-    @Override
-    public void onBindViewHolder(@NonNull IndicatorHolder holder, final int position) {
-
-        final PictureFacer pic = pictureList.get(position);
-
-        holder.positionController.setBackgroundColor(pic.getSelected() ? Color.parseColor("#00000000") : Color.parseColor("#8c000000"));
-
-        Glide.with(pictureContx)
-                .load(pic.getPicturePath())
-                .apply(new RequestOptions().centerCrop())
-                .into(holder.image);
-
-        holder.image.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                //holder.card.setCardElevation(5);
-                pic.setSelected(true);
-                notifyDataSetChanged();
-                imageListerner.onImageIndicatorClicked(position);
-            }
-        });
-
-    }
-
-    @Override
-    public int getItemCount() {
-        return pictureList.size();
-    }
-}
