+++ app/src/main/java/com/example/mannau/SubActivity.java
@@ -1,0 +1,49 @@
+package com.example.mannau;
+
+import android.os.Bundle;
+import android.view.View;
+import android.widget.AdapterView;
+import android.widget.ListView;
+import android.widget.Toast;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+
+import java.util.ArrayList;
+
+public class SubActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.weather_window);
+
+
+        weatherAdapter adapter = new weatherAdapter((((MainActivity) MainActivity.mContext).data));
+        ListView listView2 = findViewById(R.id.weather_listView);
+        listView2.setAdapter(adapter);
+
+        // 클릭 이벤트
+
+        listView2.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+            @Override
+            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+                Weather item = (((MainActivity) MainActivity.mContext).data).get(position);
+                Toast.makeText(getApplicationContext(),
+                        "온도: " +item.getTemp() + " " + "습도: "+item.getHumidity(),Toast.LENGTH_SHORT).show();
+            }
+        });
+    }
+
+    public void Show(int num, ArrayList<Weather> _data) {
+        if (num == 1) {
+            widgetAdapter adapter = new widgetAdapter(_data);
+            ListView listView = findViewById(R.id.weather_listView);
+            listView.setAdapter(adapter);
+        } else {
+            weatherAdapter adapter = new weatherAdapter(_data);
+            ListView listView2 = findViewById(R.id.weather_listView);
+            listView2.setAdapter(adapter);
+        }
+    }
+}
