+++ app/src/main/java/com/example/mannau/utils/MarginDecoration.java
@@ -1,0 +1,28 @@
+package com.example.mannau.utils;
+
+import android.content.Context;
+import android.graphics.Rect;
+import android.view.View;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.mannau.R;
+
+/**
+ * Author CodeBoy722
+ *
+ * RecyclerView's item margin decorator
+ */
+public class MarginDecoration extends RecyclerView.ItemDecoration {
+    private int margin;
+
+    public MarginDecoration(Context context) {
+        margin = context.getResources().getDimensionPixelSize(R.dimen.item_margin);
+    }
+
+    @Override
+    public void getItemOffsets(@NonNull Rect outRect, @NonNull View view, @NonNull RecyclerView parent, @NonNull RecyclerView.State state) {
+        outRect.set(margin, margin, margin, margin);
+    }
+}
