+++ app/src/main/java/com/example/mannau/utils/PicHolder.java
@@ -1,0 +1,20 @@
+package com.example.mannau.utils;
+
+import android.view.View;
+import android.widget.ImageView;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.mannau.R;
+
+
+public class PicHolder extends RecyclerView.ViewHolder {
+    public ImageView picture;
+
+    PicHolder(@NonNull View itemView) {
+        super(itemView);
+
+        picture = itemView.findViewById(R.id.image);
+    }
+}
