+++ app/src/main/res/layout/activity_image_display.xml
@@ -31,7 +31,7 @@
 
         </androidx.cardview.widget.CardView>
 
-        <com.example.immersioncamp1.utils.AutoFitRecyclerView
+        <com.example.mannau.utils.AutoFitRecyclerView
             android:id="@+id/recycler"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
@@ -46,7 +46,7 @@
             app:fastScrollVerticalThumbDrawable="@drawable/thumb_drawable"
             app:fastScrollVerticalTrackDrawable="@drawable/line_drawable"-->
 
-        </com.example.immersioncamp1.utils.AutoFitRecyclerView>
+        </com.example.mannau.utils.AutoFitRecyclerView>
 
     </RelativeLayout>
