+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">ImmersionCamp1</string>
+    <string name="app_name">MannaU</string>
     <string name="name">Name</string>
     <string name="phone_number">Phone Number</string>
     <string name="organization">Organization</string>
